Adding projects for the plug-ins. Currently incomplete.
git-svn-id: svn://svn.code.sf.net/p/playerpro/code/branches/PlayerPRO6@217 f4605116-7794-4d0b-9e48-65ca9bfeedfd

+++ MADDriver.source/MADDriver.Source/RDriverInt.h
@@ -331,7 +331,7 @@ struct __MADDriverRec
 	Boolean					newData[ MAXDRAW];
 	Boolean					useOsciBuffers;
 	short					curDrawPtr;
-	UInt32					curTime;
+	unsigned long			curTime;
 	Boolean					XMLinear, MODMode, JumpToNextPattern, endPattern, MADPlay;
 	SInt32					ASCBUFFER;
 	SInt32					BufSize;
